rename oss to slidehub
