--- lib/oss/version.rb
@@ -1,3 +1,0 @@
-module SlideHub
-  VERSION = '1.0.7'.freeze
-end
