+++ lib/slide_hub/version.rb
@@ -1,0 +1,3 @@
+module SlideHub
+  VERSION = '1.0.8'.freeze
+end
